+++ ChangeLog.md
@@ -97,6 +97,8 @@
 
   * Fix course element reloading on sync (Nathan Walters).
 
+  * Fix course element file loading (Matt West).
+
 * __2.10.1__ - 2017-05-24
 
   * Fix display of saved submissions for Exam assessments.